Refactor: clean monster related code and create a specific class for basic monster zako.
