+++ Assets/Prefab/Monster.prefab
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1e70dbd17e8c85fb98f52f8af9c90324d761320130e19b0a107c19446092a2a5
+oid sha256:95030ebe6860391bdeb4b993ef0bfbd27b93b8519d4e8b19e0fdb335fb4fdecb
 size 8980
