+++ Assets/Scenes/Experiment_MonsterScene.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:800772b8d080526a16b0e559efe7171f28ce1b71fb8be7c2b67551a1eb6218e4
-size 73673
+oid sha256:4baa77b1302f991048249035195c00bed79ffdf3f21e4e290ea2c1332267db99
+size 73320
