+++ Assets/Scripts/Monster/Zako.cs.meta
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:99fd3f6cf17a5b9899900c737aa9fbad1fbfdb225cf41a92f89cecc6e5718447
+size 243
